Web: Rename the MediaViewerTS folder to MediaViewerV2
